Move random links and login info and other stuff into the navbar.
Add two new menus in navbar for Settings and User. Needs more consolidation
and some thought into where things go in the menus.

Also needs the logic from left.cgi and right.cgi to be moved into index.cgi,
but, I'm woking with a plain HTML index.html mockup for now just to put
everything in the right places, so we find a usable solution before writing
a bunch of code.